requirements.txt
oracledb
python-dotenv
requests
urllib3
